requirements.txt
colorama==0.4.4
pyyaml==5.4.1